S05P21D204/ISEAU/backend/requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.3
certifi==2021.5.30
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-appconf==1.0.4
django-bootstrap-v5==1.0.0
django-cors-headers==3.8.0
django-extensions==3.1.1
django-imagekit==4.0.2
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
entrypoints==0.3
Faker==8.1.2
flex==6.14.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==2.1.1
inflection==0.5.1
ipykernel==5.4.3
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
lxml==4.6.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mistune==0.8.4
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
pilkit==2.0
Pillow==8.3.2
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.18.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.5.0
six==1.16.0
soupsieve==2.2
sqlparse==0.4.2
strict-rfc3339==0.7
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.4.1
haversine==2.5.1
